BoostedCDE

Documentation for BoostedCDE.

BoostedCDE.BaseLearner

BoostedCDE.BoostingModel

BoostedCDE.ConstBaseLearner

BoostedCDE.gaussian_simulator

BoostedCDE.mvn_loss

BoostedCDE.predict

BoostedCDE.triangular_to_vec

BoostedCDE.vec_to_triangular

BoostedCDE.μ_chol_splitter

BoostedCDE.@loopify

BoostedCDE.BaseLearner — Type

Base learners are used to predict the negative gradient vector. These must all have the methods fit! and predict.

BoostedCDE.BoostingModel — Type

Constructs a BoostingModel that can be trained using [boost]@ref

BoostedCDE.ConstBaseLearner — Type

Base learner used at initialization to predict a constant value. Additional arguments can be provided to fit! for consistency with other base learners but these are ignored.

BoostedCDE.gaussian_simulator — Method

Simulate a three dimensional Gaussian mean vector θ. The covariance is diagonal, and fixed to σ=0.1. Parameter vector θ is the mean vector of the Gaussian.

BoostedCDE.mvn_loss — Method

Negative log-probability of x using vector of parameters ϕ to parameterise the means and cholesky decomposition of the normal distribution.

BoostedCDE.predict — Method

Predict using the boosting model to get a vector of distributional parameters. Computes using no cached results (e.g. from previous iterations during training).

BoostedCDE.triangular_to_vec — Method

Convert triangular matrix to vector. This uses the convention that the upper triangular elements are listed columnwise (or equivilently, lower triangular elements listed rowwise).

BoostedCDE.vec_to_triangular — Method

Constructs an upper triangular matrix from a vector. Vector should correspond to the upper triangular elements listed columnwise.

BoostedCDE.μ_chol_splitter — Method

Take a flattened mean and traingular matrix and reconstruct it returning a tuple

BoostedCDE.@loopify — Macro

Convenience macro to define a method for a simulator that allows simulating in batches. i.e. it takes simulator(rng::AbstractRNG, θ::AbstractVector{Float64}) and wraps it in a for loop to get simulator(rng::AbstractRNG, θ::Matrix{Float64}).