BoostedCDE

Documentation for BoostedCDE.

BoostedCDE.BaseLearner

BoostedCDE.BoostingModel

BoostedCDE.MeanCholeskyMvn

BoostedCDE.PolyBaseLearner

BoostedCDE.boost!

BoostedCDE.fit!

BoostedCDE.gaussian_simulator

BoostedCDE.get_params

BoostedCDE.linear_θ_to_ϕ_mvn_simulator

BoostedCDE.loss

BoostedCDE.predict

BoostedCDE.predict

BoostedCDE.reset!

BoostedCDE.step!

BoostedCDE.tri_n_el_to_d

BoostedCDE.unvectorize

BoostedCDE.vectorize

BoostedCDE.@loopify

BoostedCDE.BaseLearner — Type

Base learners are used to predict the negative gradient vector. These must all have the methods fit! and predict. Note that the same model is used multiple, so it must be safe to call the fit! method multiple times with different data.

BoostedCDE.BoostingModel — Type

Constructs a BoostingModel that can be trained using boost!

BoostedCDE.MeanCholeskyMvn — Type

Struct denoting we wish to use a mean vector and cholesky decomposition of the covariance matrix to parameterize a multivariate normal distribution of dimension d.

BoostedCDE.PolyBaseLearner — Type

Base learner using polynomial transform. By defualt, with use_cache = true, the model will cache the feature vectors along with QR decompositions of the polynomial transforms to avoid recalculation. This will speed up training at the cost of increased memory usage. An IdDict is used as the cache.

BoostedCDE.boost! — Method

Fit a boosting model by performing M steps. Returns the model, predictions and losses from each training iteration.

boost!(model, θ, x; loss, steps)

BoostedCDE.fit! — Method

Fit the Base learner model to the negative gradient vector uⱼ using feature vector θ.

fit!(base_learner, θ, u)

BoostedCDE.gaussian_simulator — Method

Simulate a three dimensional Gaussian mean vector θ. The covariance is diagonal, and fixed to σ=0.1. Parameter vector θ is the mean vector of the Gaussian.

BoostedCDE.get_params — Method

Get the parameters for the distribution from the vectorised form.

BoostedCDE.linear_θ_to_ϕ_mvn_simulator — Method

Simulator that applies simple linear transformation of θ to get ϕ. Useful for testing.

BoostedCDE.loss — Method

Calculate the loss given a parameterisation specified by Abstractϕ. If matrices are used, reduction is carried out using the mean.

BoostedCDE.predict — Method

Predict using the boosting model to get the conditional distributional parameters. During training, the ϕ matrix from previous iteration can be provided to avoid recalculating.

BoostedCDE.predict — Method

Predict the negative gradient vector using θ.

predict(base_learner, θ)

BoostedCDE.reset! — Method

"Reset" the boosting model, removing all the selected base learners and corresponding indices.

BoostedCDE.step! — Method

Step the boosting model, by adding on a single new base model that minimizes the inner loss. The gradient! function should take ϕ and x and return the gradient matrix matching the shape of ϕ. Note gradient! may (or may not) be mutating depending on the definition (e.g. if using tapes with ReverseDiff it would mutate the tape).

step!(model, θ, x, ϕₘ; loss)

BoostedCDE.tri_n_el_to_d — Method

Given the number of triangular elements in a matrix, get the dimension.

BoostedCDE.unvectorize — Method

Given the array type and the vector, reform the array. This does the oposite transformation of vectorize.

BoostedCDE.vectorize — Method

Flatten to vector, with methods for special matrices, that ignore unwanted elements e.g. for diagonal matrices this will ignore off diagonal elements. These methods may return a copy or a view (behaviour is not kept consistent).

BoostedCDE.@loopify — Macro

Convenience macro to define a method for a simulator that allows simulating in batches. i.e. it takes simulator(rng::AbstractRNG, θ::AbstractVector{Float64}) and wraps it in a for loop to get simulator(rng::AbstractRNG, θ::Matrix{Float64}).