BoostedCDE

Documentation for BoostedCDE.

BoostedCDE.BaseLearner

BoostedCDE.BoostingModel

BoostedCDE.MeanCholeskyMvn

BoostedCDE.PolyBaseLearner

BoostedCDE.StandardScaler

BoostedCDE.StandardScaler

BoostedCDE.boost!

BoostedCDE.boostcv!

BoostedCDE.cost

BoostedCDE.fit!

BoostedCDE.gaussian_simulator

BoostedCDE.get_params

BoostedCDE.get_tape_∇

BoostedCDE.linear_x_to_ϕ_mvn_simulator

BoostedCDE.predict

BoostedCDE.predict

BoostedCDE.predict

BoostedCDE.reset!

BoostedCDE.step!

BoostedCDE.step_naive!

BoostedCDE.train_val_split

BoostedCDE.tri_n_el_to_d

BoostedCDE.unscale

BoostedCDE.unvectorize

BoostedCDE.vectorize

BoostedCDE.@loopify

BoostedCDE.BaseLearner — Type

Base learners are used to predict the negative gradient vector. These must all have the methods fit! and predict. Note that the same model is used multiple, so it must be safe to call the fit! method multiple times with different data.

BoostedCDE.BoostingModel — Type

Constructs a BoostingModel that can be trained using boost!

BoostedCDE.MeanCholeskyMvn — Type

Struct denoting we wish to use a mean vector and cholesky decomposition of the covariance matrix to parameterize a multivariate normal distribution of dimension d.

BoostedCDE.PolyBaseLearner — Type

Base learner using polynomial transform. By defualt, with use_cache = true, the model will cache the feature vectors along with QR decompositions of the polynomial transforms to avoid recalculation. This will speed up training at the cost of increased memory usage. An IdDict is used as the cache.

BoostedCDE.StandardScaler — Type

Create a scaler, fitted to matrix y. Can be used to apply (y - μ)/σ to each column of a matrix. Use as callable struct to scale and use unscale to unscale.

BoostedCDE.StandardScaler — Method

Transform a matrix using the scaler

BoostedCDE.boost! — Method

Minimal function to fit a boosting model. loss should take ϕ and y as arguments, and ∇loss should just take ϕ.

boost!(model, x, y; steps, loss, ∇loss, step!)

BoostedCDE.boostcv! — Method

Boosting with cross validation and patience. loss should take ϕ and y, returning a scalar. ∇loss should take ϕ (i.e. y_train should be abstracted away), returning matrix with size matching ϕ. data should be an object that can be unpacked/destructured to (xtrain, xval, ytrain, yval), e.g. NamedTuple resulting from train_val_split. Returns a NamedTuple of results.

If ∇loss is not provided, by default, the gradient is found using ReverseDiff, using ReverseDiff.GradientTape.

boostcv!(model, data; steps, loss, ∇loss, step!, max_patience)

BoostedCDE.cost — Method

Calculate the loss given a parameterisation specified by Abstractϕ. If matrices are used, reduction is carried out using summation.

BoostedCDE.fit! — Method

Fit the Base learner model to the negative gradient vector uⱼ using feature vector x.

fit!(base_learner, x, u)

BoostedCDE.gaussian_simulator — Method

Simulate a three dimensional Gaussian mean vector x. The covariance is diagonal, and fixed to σ=0.1. Parameter vector x is the mean vector of the Gaussian.

BoostedCDE.get_params — Method

Get the parameters for the distribution from the vectorised form.

BoostedCDE.get_tape_∇ — Method

Gradient function with tape, specialised on x and y (usually training data).

BoostedCDE.linear_x_to_ϕ_mvn_simulator — Method

Simulator that applies simple linear transformation of x to get ϕ. Useful for testing.

BoostedCDE.predict — Method

Predict using the boosting model to get the conditional distributional parameters, using the ϕ matrix from previous training iteration to avoid recalculating.

predict(model, x, last_ϕ)

BoostedCDE.predict — Method

Predict using the boosting model to get the conditional distributional parameters.

predict(model, x)

BoostedCDE.predict — Method

Predict the negative gradient vector using x.

predict(base_learner, x)

BoostedCDE.reset! — Method

"Reset" the boosting model, removing all the selected base learners and corresponding indices.

BoostedCDE.step! — Method

Step the boosting model, by adding on a single new base model that minimizes the gradient norm explained. u is the gradient matrix with shape matching ϕ, i.e. N×J where N is the number observations/simulations, and J is the number of distributional parameters.

step!(model, x, u)

BoostedCDE.step_naive! — Method

As for step!, but without skipping training base models where the norm explained cannot be improved.

BoostedCDE.train_val_split — Function

Train test split for two matrices. Returns named tuple with keys [xtrain, xval, ytrain, yval]

BoostedCDE.tri_n_el_to_d — Method

Given the number of triangular elements in a matrix, get the dimension.

BoostedCDE.unscale — Method

Unscale the matrix.

BoostedCDE.unvectorize — Method

Given the array type and the vector, reform the array. This does the oposite transformation of vectorize.

BoostedCDE.vectorize — Method

Flatten to vector, with methods for special matrices, that ignore unwanted elements e.g. for diagonal matrices this will ignore off diagonal elements. These methods may return a copy or a view (behaviour is not kept consistent).

BoostedCDE.@loopify — Macro

Convenience macro to define a method for a simulator that allows simulating in batches. i.e. it takes simulator(rng::AbstractRNG, x::AbstractVector{Float64}) and wraps it in a for loop to get simulator(rng::AbstractRNG, x::Matrix{Float64}).